rpcinfo: change order of version to be tried to 4, 3, 2
When specifying TCP as transport on rpcinfo getport command (-T tcp), the
initial RPC getport packet is still sent as a UDP packet.
This is due to rpc protocol version 2 is tried first and function getpmaphandle()
states:

/*
* Try UDP only - there are some portmappers out
* there that use UDP only.
*/

Even on systems with newer rpc versions, this will hang when the UDP port is blocked.
That is a quite artifical limitation, because UDP only portmappert should be really
rare by now.
The attached code changes the order of versions to be tried to 4, then 3 and
use version 2 only as fallback.

Signed-off-by: Thomas Blume <Thomas.Blume@suse.com>